Refactor project details, extract data fetching into custom hook
